backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
requests
httpx
supabase
